ready

+++ controllers/contactsController.js
@@ -4,7 +4,10 @@ const Contact = require("../models/Contact");
 
 const getAllContacts = async (req, res, next) => {
   try {
-    const result = await Contact.find();
+    const {_id: owner} = req.user;
+    const {page = 1, limit = 10} = req.query;
+    const skip = (page - 1)*limit;
+    const result = await Contact.find({owner}, '-createdAt', {skip, limit});
     res.json(result);
   } catch (error) {
     next(error);
@@ -14,7 +17,8 @@ const getAllContacts = async (req, res, next) => {
 const getById = async (req, res, next) => {
   try {
     const { contactId } = req.params;
-    const result = await Contact.findById(contactId);
+    const {_id: owner} = req.user;
+    const result = await Contact.findOne({_id: contactId, owner});
     if (!result) {
       throw helpers.HttpError(404, `Not found`);
     }
@@ -29,7 +33,8 @@ const add = async (req, res, next) => {
     if (error) {
       throw helpers.HttpError(400, error.message);
     }
-    const result = await Contact.create(req.body);
+    const {_id: owner} = req.user;
+    const result = await Contact.create({...req.body, owner});
     res.status(201).json(result);
   } catch (error) {
     next(error);
@@ -39,7 +44,8 @@ const add = async (req, res, next) => {
 const deleteContact = async (req, res, next) => {
   try {
     const { contactId } = req.params;
-    const result = await Contact.findByIdAndDelete(contactId);
+    const {_id: owner} = req.user;
+    const result = await Contact.findByIdAndDelete({_id: contactId, owner});
     if (!result) {
       throw helpers.HttpError(404, `Not found`);
     }
@@ -55,7 +61,8 @@ const updateContact = async (req, res, next) => {
       throw helpers.HttpError(400, error.message);
     }
     const { contactId } = req.params;
-    const result = await Contact.findByIdAndUpdate(contactId, req.body);
+    const {_id: owner} = req.user;
+    const result = await Contact.findByIdAndUpdate({_id: contactId, owner}, req.body);
     if (!result) {
       throw helpers.HttpError(404, `Not found`);
     }
